# scripts/openapi.schema.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "LLM Guardrails Gateway",
    "description": "Safety middleware for LLM traffic and agent skills \u2014 gateway, policy, and leak scanning.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/policy": {
      "get": {
        "tags": [
          "Policy"
        ],
        "summary": "Get Policy",
        "operationId": "get_policy_policy_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PolicyOut"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "patch": {
        "tags": [
          "Policy"
        ],
        "summary": "Update Policy",
        "operationId": "update_policy_policy_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PolicyUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PolicyOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/policy/reset": {
      "post": {
        "tags": [
          "Policy"
        ],
        "summary": "Reset Policy",
        "description": "Restore the org's policy to the system defaults.",
        "operationId": "reset_policy_policy_reset_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PolicyOut"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/policy/diff": {
      "post": {
        "tags": [
          "Policy"
        ],
        "summary": "Diff Policy",
        "description": "Compare two policy blobs field by field. No DB write \u2014 pure comparison.\n\nNested structures inside the rule blobs (e.g. pii_patterns entries) are\ncompared as whole values per top-level field, not recursively.",
        "operationId": "diff_policy_policy_diff_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PolicyDiffRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PolicyDiffEntry"
                  },
                  "type": "array",
                  "title": "Response Diff Policy Policy Diff Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PolicyDiffEntry": {
        "properties": {
          "field": {
            "type": "string",
            "title": "Field"
          },
          "before": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Before"
          },
          "after": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "After"
          }
        },
        "type": "object",
        "required": [
          "field"
        ],
        "title": "PolicyDiffEntry"
      },
      "PolicyDiffRequest": {
        "properties": {
          "policy_a": {
            "additionalProperties": true,
            "type": "object",
            "title": "Policy A"
          },
          "policy_b": {
            "additionalProperties": true,
            "type": "object",
            "title": "Policy B"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "policy_a",
          "policy_b"
        ],
        "title": "PolicyDiffRequest"
      },
      "PolicyOut": {
        "properties": {
          "input_rules": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Input Rules"
          },
          "output_rules": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Output Rules"
          },
          "topic_policy": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Topic Policy"
          },
          "compliance_rules": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Compliance Rules"
          },
          "llm_backend": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Llm Backend"
          },
          "llm_model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Llm Model"
          },
          "rate_limit_rpm": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rate Limit Rpm"
          },
          "rate_limit_rpd": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rate Limit Rpd"
          },
          "org_id": {
            "type": "string",
            "title": "Org Id"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "org_id",
          "updated_at"
        ],
        "title": "PolicyOut"
      },
      "PolicyUpdate": {
        "properties": {
          "input_rules": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Input Rules"
          },
          "output_rules": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Output Rules"
          },
          "topic_policy": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Topic Policy"
          },
          "compliance_rules": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Compliance Rules"
          },
          "llm_backend": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Llm Backend"
          },
          "llm_model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Llm Model"
          },
          "rate_limit_rpm": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rate Limit Rpm"
          },
          "rate_limit_rpd": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rate Limit Rpd"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "PolicyUpdate",
        "description": "Partial update \u2014 only send what you want to change."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}